Move ParentChildHelper to test sources
